refactor: moved hooks to folders
