+++ apps/davi/src/stores/modules/guilds/SnapshotERC20Guild/fetchers/rpc/useMemberCount/index.ts
@@ -1,0 +1,1 @@
+export { useMemberCount } from './useMemberCount';
