+++ apps/davi/src/stores/modules/guilds/SnapshotERC20Guild/fetchers/rpc/useTotalLocked/useTotalLocked.ts
@@ -1,6 +1,6 @@
 import { useContractRead } from 'wagmi';
 import { BigNumber } from 'ethers';
-import { useListenToLockAndWithdrawTokens } from '../../events/useListenToLockAndWithdrawTokens';
+import { useListenToLockAndWithdrawTokens } from '../../../events/useListenToLockAndWithdrawTokens';
 import { SnapshotERC20Guild } from 'contracts/ts-files/SnapshotERC20Guild';
 import { useSnapshotId } from 'stores/modules/guilds/common/fetchers/rpc';
 import { FetcherHooksInterface } from 'stores/types';
