+++ apps/davi/src/stores/modules/guilds/SnapshotERC20Guild/fetchers/rpc/useVoterLockTimestamp/useVoterLockTimestamp.ts
@@ -1,7 +1,7 @@
 import { unix } from 'moment';
 import { useContractRead } from 'wagmi';
 import { BaseERC20Guild } from 'contracts/ts-files/BaseERC20Guild';
-import { useListenToLockAndWithdrawTokens } from '../../events/useListenToLockAndWithdrawTokens';
+import { useListenToLockAndWithdrawTokens } from '../../../events/useListenToLockAndWithdrawTokens';
 import { useListenToVoteAdded } from 'stores/modules/guilds/common/events';
 import { FetcherHooksInterface } from 'stores/types';
