+++ apps/davi/src/stores/modules/guilds/SnapshotRepERC20Guild/fetchers/rpc/useTotalLocked/index.ts
@@ -1,0 +1,1 @@
+export { useTotalLocked } from './useTotalLocked';
