+++ apps/davi/src/stores/modules/guilds/SnapshotRepERC20Guild/fetchers/rpc/useTotalLocked/useTotalLocked.ts
@@ -2,7 +2,7 @@ import { BigNumber } from 'ethers';
 import { useContractRead } from 'wagmi';
 import { BaseERC20Guild } from 'contracts/ts-files/BaseERC20Guild';
 import { ERC20SnapshotRep } from 'contracts/ts-files/ERC20SnapshotRep';
-import { useListenToTokenTransfer } from '../../events/useListenToTokenTransfer';
+import { useListenToTokenTransfer } from '../../../events/useListenToTokenTransfer';
 import {
   useDAOToken,
   useSnapshotId,
