+++ apps/davi/src/stores/modules/guilds/SnapshotRepERC20Guild/fetchers/rpc/useVoterLockTimestamp/index.ts
@@ -1,0 +1,1 @@
+export { useVoterLockTimestamp } from './useVoterLockTimestamp';
