+++ apps/davi/src/stores/modules/guilds/SnapshotRepERC20Guild/fetchers/rpc/useVotingPowerOf/index.ts
@@ -1,0 +1,1 @@
+export { useVotingPowerOf } from './useVotingPowerOf';
