+++ apps/davi/src/stores/modules/guilds/SnapshotRepERC20Guild/fetchers/rpc/useVotingPowerOf/useVotingPowerOf.test.ts
